requirements.txt
streamlit
torch
torchvision
transformers
pillow
opencv-python
numpy
tqdm
requests
